backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pydantic
openai
pypdf
python-docx
crawl4ai
python-dotenv
json-repair